refactor(Dataource\AbstractCronTask): move carbon intensity specific code
