+++ src/DataSource/CarbonIntensity/ElectricityMaps/CronTask.php
@@ -34,7 +34,7 @@ namespace GlpiPlugin\Carbon\DataSource\CarbonIntensity\ElectricityMaps;
 
 use CommonGLPI;
 use CronTask as GlpiCronTask;
-use GlpiPlugin\Carbon\DataSource\AbstractCronTask;
+use GlpiPlugin\Carbon\DataSource\CarbonIntensity\AbstractCronTask;
 use GlpiPlugin\Carbon\DataSource\CronTaskInterface;
 
 class CronTask extends AbstractCronTask implements CronTaskInterface
